% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{env.matrices}
\alias{env.matrices}
\title{Environment Matrices}
\usage{
env.matrices(env)
}
\arguments{
\item{env}{a character string vector containing the environments.}
}
\value{
Returns a list of two dataframes. The first, shown the environment in matrix form. The second provides the frequencies of each amino acid within the environments.
}
\description{
Provides the frequencies of each amino acid within the environment.
}
\examples{
env.matrices(c('ANQRmCTPQ', 'LYPPmQTPC', 'XXGSmSGXX'))
}
\references{
Aledo et al. Sci Rep. 2015; 5: 16955. (PMID: 26597773)
}
\seealso{
env.extract(), env.Ztest() and env.plot()
}
\author{
Juan Carlos Aledo
}
