% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metosite.R
\name{meto.list}
\alias{meto.list}
\title{List Proteins Found in MetOSite Matching a Keyword}
\usage{
meto.list(keyword)
}
\arguments{
\item{keyword}{a character string corresponding to the keyword}
}
\value{
This function returns a dataframe with the uniprot id, the protein name and the species, for those proteins present into MetOSite whose name contains the keyword.
}
\description{
Lists proteins found in MetOSite with names matching the keyword.
}
\examples{
meto.list('inhibitor')
}
\references{
Valverde et al. 2019. Bioinformatics 35:4849-4850 (PMID: 31197322)
}
\seealso{
meto.search(), meto.scan()
}
\author{
Juan Carlos Aledo
}
