% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa.R
\name{renum.meto}
\alias{renum.meto}
\title{Renumerate Residue Position}
\usage{
renum.meto(uniprot)
}
\arguments{
\item{uniprot}{the UniProt ID.}
}
\value{
Returns a dataframe containing the re-numerated sequence.
}
\description{
Renumerates residue position of a MetOSite sequence to match the corresponding UniProt sequence
}
\examples{
renum.meto('P01009')
}
\seealso{
is.at(), aa.at(), renum.pdb(), renum(), aa.comp()
}
\author{
Juan Carlos Aledo
}
